requirements.txt
flask
mysql-connector-python
werkzeug
python-dotenv
PyPDF2
python-docx